reuse update_device and move to mixins
