+++ homeassistant/components/mqtt/models.py
@@ -3,9 +3,12 @@ from __future__ import annotations
 
 from collections.abc import Awaitable, Callable
 import datetime as dt
-from typing import Union
+from typing import TypedDict, Union
 
 import attr
+
+from homeassistant.helpers.template import Template
+from homeassistant.helpers.typing import ConfigType
 
 PublishPayloadType = Union[str, bytes, int, float, None]
 ReceivePayloadType = Union[str, bytes]
@@ -35,3 +38,28 @@ class ReceiveMessage:
 
 AsyncMessageCallbackType = Callable[[ReceiveMessage], Awaitable[None]]
 MessageCallbackType = Callable[[ReceiveMessage], None]
+
+
+class MqttNotificationConfig(TypedDict, total=False):
+    """Supply service parameters for MqttNotificationService."""
+
+    command_topic: str
+    command_template: Template
+    encoding: str
+    name: str
+    qos: int
+    retain: bool
+    targets: list
+    title: str
+    device: ConfigType
+
+
+class MqttTagConfig(TypedDict, total=False):
+    """Supply service parameters for MQTTTagScanner."""
+
+    topic: str
+    value_template: Template
+    device: ConfigType
+
+
+MqttTypedDictConfigType = Union[MqttNotificationConfig, MqttTagConfig]
